Refactors pvc code in occlient and kclient
