+++ pkg/devfile/adapters/kubernetes/storage/utils.go
@@ -99,7 +99,7 @@ func GetExistingPVC(Client *kclient.Client, volumeName, componentName string) (s
 
 	klog.V(2).Infof("Checking PVC for volume %v and label %v\n", volumeName, label)
 
-	PVCs, err := Client.GetPVCsFromSelector(label)
+	PVCs, err := Client.ListPVCs(label)
 	if err != nil {
 		return "", errors.Wrapf(err, "Unable to get PVC with selectors "+label)
 	}
@@ -118,7 +118,7 @@ func GetExistingPVC(Client *kclient.Client, volumeName, componentName string) (s
 // DeleteOldPVCs deletes all the old PVCs which are not in the processedVolumes map
 func DeleteOldPVCs(Client *kclient.Client, componentName string, processedVolumes map[string]bool) error {
 	label := fmt.Sprintf("component=%s", componentName)
-	PVCs, err := Client.GetPVCsFromSelector(label)
+	PVCs, err := Client.ListPVCs(label)
 	if err != nil {
 		return errors.Wrapf(err, "unable to get PVC with selectors "+label)
 	}
@@ -134,4 +134,76 @@ func DeleteOldPVCs(Client *kclient.Client, componentName string, processedVolume
 		}
 	}
 	return nil
+}
+
+// GetPVCAndVolumeMount gets the PVC and updates the containers with the volume mount
+// volumeNameToPVCName is a map of volume name to the PVC created
+// containerNameToVolumes is a map of the Devfile container names to the Devfile Volumes
+func GetPVCAndVolumeMount(containers []corev1.Container, volumeNameToPVCName map[string]string, containerNameToVolumes map[string][]common.DevfileVolume) ([]corev1.Container, []corev1.Volume, error) {
+	var pvcVols []corev1.Volume
+	for volName, pvcName := range volumeNameToPVCName {
+		generatedVolumeName, err := generateVolumeNameFromPVC(pvcName)
+		if err != nil {
+			return nil, nil, errors.Wrapf(err, "Unable to generate volume name from pvc name")
+		}
+		pvcVols = append(pvcVols, getPVC(generatedVolumeName, pvcName))
+
+		// containerNameToMountPaths is a map of the Devfile container name to their Devfile Volume Mount Paths for a given Volume Name
+		containerNameToMountPaths := make(map[string][]string)
+		for containerName, volumes := range containerNameToVolumes {
+			for _, volume := range volumes {
+				if volName == volume.Name {
+					containerNameToMountPaths[containerName] = append(containerNameToMountPaths[containerName], volume.ContainerPath)
+				}
+			}
+		}
+
+		containers = addVolumeMountToContainers(containers, generatedVolumeName, containerNameToMountPaths)
+	}
+	return containers, pvcVols, nil
+}
+
+// getPVC gets a pvc type volume with the given volume name and pvc name
+func getPVC(volumeName, pvcName string) corev1.Volume {
+
+	return corev1.Volume{
+		Name: volumeName,
+		VolumeSource: corev1.VolumeSource{
+			PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
+				ClaimName: pvcName,
+			},
+		},
+	}
+}
+
+// generateVolumeNameFromPVC generates a volume name based on the pvc name
+func generateVolumeNameFromPVC(pvc string) (volumeName string, err error) {
+	volumeName, err = util.NamespaceOpenShiftObject(pvc, "vol")
+	if err != nil {
+		return "", err
+	}
+	return
+}
+
+// addVolumeMountToContainers adds the Volume Mounts in containerNameToMountPaths to the containers for a given pvc and volumeName
+// containerNameToMountPaths is a map of a container name to an array of its Mount Paths
+func addVolumeMountToContainers(containers []corev1.Container, volumeName string, containerNameToMountPaths map[string][]string) []corev1.Container {
+
+	for containerName, mountPaths := range containerNameToMountPaths {
+		for i, container := range containers {
+			if container.Name == containerName {
+				for _, mountPath := range mountPaths {
+					container.VolumeMounts = append(container.VolumeMounts, corev1.VolumeMount{
+						Name:      volumeName,
+						MountPath: mountPath,
+						SubPath:   "",
+					},
+					)
+				}
+				containers[i] = container
+			}
+		}
+	}
+
+	return containers
 }
