+++ pkg/kclient/volumes.go
@@ -1,8 +1,6 @@
 package kclient
 
 import (
-	"github.com/openshift/odo/pkg/devfile/adapters/common"
-	"github.com/openshift/odo/pkg/util"
 	"github.com/pkg/errors"
 
 	corev1 "k8s.io/api/core/v1"
@@ -35,71 +33,8 @@ func (c *Client) DeletePVC(pvcName string) error {
 	return c.KubeClient.CoreV1().PersistentVolumeClaims(c.Namespace).Delete(pvcName, &metav1.DeleteOptions{})
 }
 
-// GetPVC gets a pvc type volume with the given volume name and pvc name
-func GetPVC(volumeName, pvcName string) corev1.Volume {
-
-	return corev1.Volume{
-		Name: volumeName,
-		VolumeSource: corev1.VolumeSource{
-			PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-				ClaimName: pvcName,
-			},
-		},
-	}
-}
-
-// AddVolumeMountToContainers adds the Volume Mounts in containerNameToMountPaths to the containers for a given pvc and volumeName
-// containerNameToMountPaths is a map of a container name to an array of its Mount Paths
-func AddVolumeMountToContainers(containers []corev1.Container, volumeName string, containerNameToMountPaths map[string][]string) []corev1.Container {
-
-	for containerName, mountPaths := range containerNameToMountPaths {
-		for i, container := range containers {
-			if container.Name == containerName {
-				for _, mountPath := range mountPaths {
-					container.VolumeMounts = append(container.VolumeMounts, corev1.VolumeMount{
-						Name:      volumeName,
-						MountPath: mountPath,
-						SubPath:   "",
-					},
-					)
-				}
-				containers[i] = container
-			}
-		}
-	}
-
-	return containers
-}
-
-// GetPVCAndVolumeMount gets the PVC and updates the containers with the volume mount
-// volumeNameToPVCName is a map of volume name to the PVC created
-// containerNameToVolumes is a map of the Devfile container names to the Devfile Volumes
-func GetPVCAndVolumeMount(containers []corev1.Container, volumeNameToPVCName map[string]string, containerNameToVolumes map[string][]common.DevfileVolume) ([]corev1.Container, []corev1.Volume, error) {
-	var pvcVols []corev1.Volume
-	for volName, pvcName := range volumeNameToPVCName {
-		generatedVolumeName, err := generateVolumeNameFromPVC(pvcName)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "Unable to generate volume name from pvc name")
-		}
-		pvcVols = append(pvcVols, GetPVC(generatedVolumeName, pvcName))
-
-		// containerNameToMountPaths is a map of the Devfile container name to their Devfile Volume Mount Paths for a given Volume Name
-		containerNameToMountPaths := make(map[string][]string)
-		for containerName, volumes := range containerNameToVolumes {
-			for _, volume := range volumes {
-				if volName == volume.Name {
-					containerNameToMountPaths[containerName] = append(containerNameToMountPaths[containerName], volume.ContainerPath)
-				}
-			}
-		}
-
-		containers = AddVolumeMountToContainers(containers, generatedVolumeName, containerNameToMountPaths)
-	}
-	return containers, pvcVols, nil
-}
-
-// GetPVCsFromSelector returns the PVCs based on the given selector
-func (c *Client) GetPVCsFromSelector(selector string) ([]corev1.PersistentVolumeClaim, error) {
+// ListPVCs returns the PVCs based on the given selector
+func (c *Client) ListPVCs(selector string) ([]corev1.PersistentVolumeClaim, error) {
 	pvcList, err := c.KubeClient.CoreV1().PersistentVolumeClaims(c.Namespace).List(metav1.ListOptions{
 		LabelSelector: selector,
 	})
@@ -110,11 +45,32 @@ func (c *Client) GetPVCsFromSelector(selector string) ([]corev1.PersistentVolume
 	return pvcList.Items, nil
 }
 
-// generateVolumeNameFromPVC generates a volume name based on the pvc name
-func generateVolumeNameFromPVC(pvc string) (volumeName string, err error) {
-	volumeName, err = util.NamespaceOpenShiftObject(pvc, "vol")
+// ListPVCNames returns the PVC names for the given selector
+func (c *Client) ListPVCNames(selector string) ([]string, error) {
+	pvcs, err := c.ListPVCs(selector)
 	if err != nil {
-		return "", err
+		return nil, errors.Wrap(err, "unable to get PVCs from selector")
 	}
-	return
+
+	var names []string
+	for _, pvc := range pvcs {
+		names = append(names, pvc.Name)
+	}
+
+	return names, nil
+}
+
+// GetPVCFromName returns the PVC of the given name
+func (c *Client) GetPVCFromName(pvcName string) (*corev1.PersistentVolumeClaim, error) {
+	return c.KubeClient.CoreV1().PersistentVolumeClaims(c.Namespace).Get(pvcName, metav1.GetOptions{})
+}
+
+// UpdatePVCLabels updates the given PVC with the given labels
+func (c *Client) UpdatePVCLabels(pvc *corev1.PersistentVolumeClaim, labels map[string]string) error {
+	pvc.Labels = labels
+	_, err := c.KubeClient.CoreV1().PersistentVolumeClaims(c.Namespace).Update(pvc)
+	if err != nil {
+		return errors.Wrap(err, "unable to remove storage label from PVC")
+	}
+	return nil
 }
