+++ pkg/storage/storage.go
@@ -82,7 +82,7 @@ func Unmount(client *occlient.Client, storageName string, componentName string, 
 		return errors.Wrapf(err, "unable to remove volume: %v from Deployment Config: %v", pvcName, dc.Name)
 	}
 
-	pvc, err := client.GetPVCFromName(pvcName)
+	pvc, err := client.GetKubeClient().GetPVCFromName(pvcName)
 	if err != nil {
 		return errors.Wrapf(err, "unable to get PersistentVolumeClaim from name: %v", pvcName)
 	}
@@ -90,7 +90,7 @@ func Unmount(client *occlient.Client, storageName string, componentName string, 
 	pvcLabels[storagelabels.StorageLabel] = storageName
 
 	if updateLabels {
-		if err := client.UpdatePVCLabels(pvc, pvcLabels); err != nil {
+		if err := client.GetKubeClient().UpdatePVCLabels(pvc, pvcLabels); err != nil {
 			return errors.Wrapf(err, "unable to remove storage label from : %v", pvc.Name)
 		}
 	}
@@ -106,7 +106,7 @@ func Delete(client *occlient.Client, name string) error {
 	}
 
 	// delete the associated PVC with the component
-	err = client.DeletePVC(pvcName)
+	err = client.GetKubeClient().DeletePVC(pvcName)
 	if err != nil {
 		return errors.Wrapf(err, "unable to delete PVC %v", pvcName)
 	}
@@ -125,7 +125,7 @@ func List(client *occlient.Client, componentName string, applicationName string)
 		return StorageList{}, errors.Wrapf(err, "unable to get Deployment Config associated with component %v", componentName)
 	}
 
-	pvcs, err := client.GetPVCsFromSelector(storagelabels.StorageLabel)
+	pvcs, err := client.GetKubeClient().ListPVCs(storagelabels.StorageLabel)
 	if err != nil {
 		return StorageList{}, errors.Wrapf(err, "unable to get PVC using selector %v", storagelabels.StorageLabel)
 	}
@@ -213,7 +213,7 @@ func ListMounted(client *occlient.Client, componentName string, applicationName 
 
 // ListUnmounted lists all the unmounted storage associated with the given application
 func ListUnmounted(client *occlient.Client, applicationName string) (StorageList, error) {
-	pvcs, err := client.GetPVCsFromSelector(storagelabels.StorageLabel)
+	pvcs, err := client.GetKubeClient().ListPVCs(storagelabels.StorageLabel)
 	if err != nil {
 		return StorageList{}, errors.Wrapf(err, "unable to get PVC using selector %v", storagelabels.StorageLabel)
 	}
@@ -246,7 +246,7 @@ func Exists(client *occlient.Client, storageName string, applicationName string)
 	labels[applabels.ApplicationLabel] = applicationName
 	labels[storagelabels.StorageLabel] = storageName
 	selector := util.ConvertLabelsToSelector(labels)
-	pvcs, err := client.GetPVCsFromSelector(selector)
+	pvcs, err := client.GetKubeClient().ListPVCs(selector)
 	if err != nil {
 		return false, errors.Wrapf(err, "unable to list storage for application %v", applicationName)
 	}
@@ -277,7 +277,7 @@ func getPVCNameFromStorageName(client *occlient.Client, storageName string) (str
 	labels[storagelabels.StorageLabel] = storageName
 
 	selector := util.ConvertLabelsToSelector(labels)
-	pvcs, err := client.GetPVCNamesFromSelector(selector)
+	pvcs, err := client.GetKubeClient().ListPVCNames(selector)
 	if err != nil {
 		return "", errors.Wrapf(err, "unable to get PVC names for selector %v", selector)
 	}
@@ -294,7 +294,7 @@ func GetComponentNameFromStorageName(client *occlient.Client, storageName string
 	labels[storagelabels.StorageLabel] = storageName
 
 	selector := util.ConvertLabelsToSelector(labels)
-	pvcs, err := client.GetPVCsFromSelector(selector)
+	pvcs, err := client.GetKubeClient().ListPVCs(selector)
 	if err != nil {
 		return "", errors.Wrap(err, "unable to list the pvcs")
 	}
@@ -341,7 +341,7 @@ func Mount(client *occlient.Client, path string, storageName string, componentNa
 		return errors.Wrapf(err, "unable to create namespaced name")
 	}
 
-	pvc, err := client.GetPVCFromName(generatePVCNameFromStorageName(namespacedOpenShiftObject))
+	pvc, err := client.GetKubeClient().GetPVCFromName(generatePVCNameFromStorageName(namespacedOpenShiftObject))
 	if err != nil {
 		return errors.Wrap(err, "unable to get the pvc from the storage name")
 	}
@@ -359,7 +359,7 @@ func Mount(client *occlient.Client, path string, storageName string, componentNa
 	if err := client.AddPVCToDeploymentConfig(dc, pvc.Name, path); err != nil {
 		return errors.Wrap(err, "unable to add PVC to DeploymentConfig")
 	}
-	err = client.UpdatePVCLabels(pvc, storagelabels.GetLabels(storageName, componentName, applicationName, true))
+	err = client.GetKubeClient().UpdatePVCLabels(pvc, storagelabels.GetLabels(storageName, componentName, applicationName, true))
 	if err != nil {
 		return errors.Wrap(err, "unable to update the pvc")
 	}
@@ -604,7 +604,7 @@ func DevfileListMounted(kClient *kclient.Client, componentName string) (StorageL
 	}
 
 	label := fmt.Sprintf("component=%s", componentName)
-	pvcs, err := kClient.GetPVCsFromSelector(label)
+	pvcs, err := kClient.ListPVCs(label)
 	if err != nil {
 		return StorageList{}, errors.Wrapf(err, "unable to get PVC using selector %v", storagelabels.StorageLabel)
 	}
